refactor : core/model running_result -> battle, single 디렉토리 분리 #PAR-340
