+++ core/data/src/main/java/online/partyrun/partyrunapplication/core/data/repository/ResultRepository.kt
@@ -1,7 +1,7 @@
 package online.partyrun.partyrunapplication.core.data.repository
 
 import kotlinx.coroutines.flow.Flow
-import online.partyrun.partyrunapplication.core.model.running_result.BattleResult
+import online.partyrun.partyrunapplication.core.model.running_result.battle.BattleResult
 import online.partyrun.partyrunapplication.core.common.result.Result
 
 interface ResultRepository {
