+++ core/domain/src/main/java/online/partyrun/partyrunapplication/core/domain/running_result/GetBattleResultUseCase.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 import online.partyrun.partyrunapplication.core.common.result.Result
 import online.partyrun.partyrunapplication.core.common.result.mapResultModel
 import online.partyrun.partyrunapplication.core.data.repository.ResultRepository
-import online.partyrun.partyrunapplication.core.model.running_result.toUiModel
+import online.partyrun.partyrunapplication.core.model.running_result.battle.toUiModel
 import online.partyrun.partyrunapplication.core.model.running_result.ui.BattleResultUiModel
 import javax.inject.Inject
