+++ core/network/src/main/java/online/partyrun/partyrunapplication/core/network/model/response/BattleResultResponse.kt
@@ -1,7 +1,7 @@
 package online.partyrun.partyrunapplication.core.network.model.response
 
 import com.google.gson.annotations.SerializedName
-import online.partyrun.partyrunapplication.core.model.running_result.BattleResult
+import online.partyrun.partyrunapplication.core.model.running_result.battle.BattleResult
 import online.partyrun.partyrunapplication.core.model.util.DateTimeUtils.localDateTimeFormatter
 import online.partyrun.partyrunapplication.core.network.model.util.formatDate
 import online.partyrun.partyrunapplication.core.network.model.util.formatDistanceInKm
