+++ core/network/src/main/java/online/partyrun/partyrunapplication/core/network/model/response/BattleRunnerRecordResponse.kt
@@ -1,7 +1,7 @@
 package online.partyrun.partyrunapplication.core.network.model.response
 
 import com.google.gson.annotations.SerializedName
-import online.partyrun.partyrunapplication.core.model.running_result.BattleRunnerRecord
+import online.partyrun.partyrunapplication.core.model.running_result.battle.BattleRunnerRecord
 import online.partyrun.partyrunapplication.core.model.util.DateTimeUtils.localDateTimeFormatter
 import java.time.LocalDateTime
